requirements.txt
psycopg2-binary==2.9.9
yfinance==0.2.32
pandas==2.1.3
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
apache-airflow-providers-google==10.10.0
apache-airflow==2.8.0
